Format types properly

+++ esphome/components/nextion/nextion_upload_idf.cpp
@@ -88,11 +88,11 @@ Nextion::TFTUploadResult Nextion::upload_by_chunks_(esp_http_client_handle_t htt
       this->content_length_ -= read_len;
 #ifdef USE_PSRAM
       ESP_LOGD(TAG, "Uploaded %0.2f %%, remaining %" PRIu32 " bytes, free heap: %" PRIu32 " (DRAM) + %" PRIu32 " (PSRAM) bytes",
-              100.0f * (this->tft_size_ - this->content_length_) / this->tft_size_, this->content_length_,
+              static_cast<double>(100.0 * (this->tft_size_ - this->content_length_) / this->tft_size_), this->content_length_,
               heap_caps_get_free_size(MALLOC_CAP_INTERNAL), heap_caps_get_free_size(MALLOC_CAP_SPIRAM));
 #else
       ESP_LOGD(TAG, "Uploaded %0.2f %%, remaining %" PRIu32 " bytes, free heap: %" PRIu32 " bytes",
-               100.0f * (this->tft_size_ - this->content_length_) / this->tft_size_, this->content_length_,
+               static_cast<double>(100.0 * (this->tft_size_ - this->content_length_) / this->tft_size_), this->content_length_,
                esp_get_free_heap_size());
 #endif
       upload_first_chunk_sent_ = true;
